player: add track-list/N/image sub-property
This exposes whether a video track is detected as an image.

The lavf image check is moved to where the number of frames is available
for comparison, and is simplified to just check if the number of frames
if 1. Unlike checking the codec name, this detects 1-frame gifs as
images, and doesn't misdetect videos with a codec commonly used for
images, like mjpeg videos, and also webp and avif videos once ffmpeg is
capable of playing them.

pix files are also now detected as images, while before they weren't
since the condition was checking if the AVInputFormat name ends with
_pipe, and alias_pix doesn't.

Both nb_frames and codec_info_nb_frames are checked because nb_frames is
0 for some video codecs (hevc, av1, vc1, mpeg1video, vp9 if forcing
--demuxer=lavf), and codec_info_nb_frames is 1 for others (mpeg, mpeg4,
wmv3). Checking both seems to always work.

libavdevice inputs like av://lavfi:testsrc have nb_frames = 0 and
codec_info_nb_frames = 1, so the condition explictly checks that streams
aren't of this type since they aren't images.

When the lavf demuxer fails, the mf one guesses if the file is an image
by its extension, so sh->image is set to true when the mf demuxer
succeds and there's only one file.

Even if you add a video's file type to --mf-type and open it with
mf://foo, only the first frame is used, so setting image to true is
still accurate.

When converting an image to the extensions listed in demux/demux_mf.c,
tga and pam files are currently the only ones detected by the mf demuxer
rather than lavf. Actually they are detected with the image2 format, but
it is blacklisted; see d0fee0ac33a.

The mkv demuxer just sets image to true for any attached picture.

The timeline demuxer just copies the value of image from source to
destination, which is imperfect since you could concatenate multiple
images (which should be done with the mf demuxer anyway), but good
enough for EDL which is based on heuristics.